Troca de packages e adequação das activities para nova novo uso das view models
